requirements.txt
Flask~=2.3.2
WTForms~=3.0.1
flask
flask-sqlalchemy
flask-migrate
ipython
python-dotenv